First Demonstration Model
Working Visuals for a setup where the algorithms are followed and there is total visbility (no walls and all points visible) (the defender does not go for the attacker, as it theoretically would if it could see it.) About to create a seperate version of the start from scratch file where the defender ignores its current thinking and just chases the attacker. (attacker will still "destroy" nodes and won't try to simply run away).